Chore[COMP345-78]: Refactor strategy
