+++ ProjectSourceCode/Strategy/characteractionstrategy.h
@@ -9,7 +9,7 @@
 #include <vector>
 #include <string>
 
-#include "Interactable.h"
+#include "..\Interactable.h"
 
 /*!
 * \namespace characteractionstrategy
