+++ ProjectSourceCode/Strategy/friendlystrategy.h
