+++ ProjectSourceCode/Strategy/humanplayerstrategy.h
