+++ [COMP345] Assignment 1.vcxproj
@@ -120,9 +120,9 @@
     <ClCompile Include="ProjectSourceCode\campaign.cpp" />
     <ClCompile Include="ProjectSourceCode\testcampaign.cpp" />
     <ClCompile Include="ProjectSourceCode\Decorator\componentdecorator.cpp" />
-    <ClCompile Include="ProjectSourceCode\humanplayerstrategy.cpp" />
-    <ClCompile Include="ProjectSourceCode\aggressorstrategy.cpp" />
-    <ClCompile Include="ProjectSourceCode\friendlystrategy.cpp" />
+    <ClCompile Include="ProjectSourceCode\Strategy\humanplayerstrategy.cpp" />
+    <ClCompile Include="ProjectSourceCode\Strategy\aggressorstrategy.cpp" />
+    <ClCompile Include="ProjectSourceCode\Strategy\friendlystrategy.cpp" />
     <!-- <ClCompile Include="ProjectSourceCode\testSerializeCharacter.cpp" /> -->
     <ClCompile Include="ProjectSourceCode\EmptyCell.cpp" />
     <ClCompile Include="ProjectSourceCode\Interactable.cpp" />
@@ -167,10 +167,10 @@
     <!-- <ClInclude Include="ProjectSourceCode\testSerializeCharacter.h" /> -->
     <ClInclude Include="ProjectSourceCode\Decorator\abstractcomponent.h" />
     <ClInclude Include="ProjectSourceCode\Decorator\componentdecorator.h" />
-    <ClInclude Include="ProjectSourceCode\characteractionstrategy.h" />
-    <ClInclude Include="ProjectSourceCode\humanplayerstrategy.h" />
-    <ClInclude Include="ProjectSourceCode\aggressorstrategy.h" />
-    <ClInclude Include="ProjectSourceCode\friendlystrategy.h" />
+    <ClInclude Include="ProjectSourceCode\Strategy\characteractionstrategy.h" />
+    <ClInclude Include="ProjectSourceCode\Strategy\humanplayerstrategy.h" />
+    <ClInclude Include="ProjectSourceCode\Strategy\aggressorstrategy.h" />
+    <ClInclude Include="ProjectSourceCode\Strategy\friendlystrategy.h" />
     <ClInclude Include="ProjectSourceCode\EmptyCell.h" />
     <ClInclude Include="ProjectSourceCode\Interactable.h" />
     <ClInclude Include="ProjectSourceCode\MapBuilder.h" />
